backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
google-auth
schedule
pdfplumber
requests
bs4
